user can add the same personal user as built-in and globals, user is able to delete flavour with username included
